\name{pakietAnity-package}
\alias{pakietAnity-package}
\alias{pakietAnity}
\docType{package}
\title{\packageTitle{pakietAnity}}
\description{\packageDescription{pakietAnity}}
\section{Package Content}{\packageIndices{pakietAnity}}
\author{\packageAuthor{pakietAnity}}
\section{Maintainer}{\packageMaintainer{pakietAnity}}
\keyword{package}
